Simplify query service
