+++ src/main/scala/com/github/scalalab3/logs/services/StorageProvider.scala
@@ -1,0 +1,7 @@
+package com.github.scalalab3.logs.services
+
+import akka.actor.ActorRef
+
+trait StorageProvider {
+  def dbService: ActorRef
+}
