--- src/main/scala/com/github/scalalab3/logs/services/messages/Request.scala
@@ -1,3 +1,0 @@
-package com.github.scalalab3.logs.services.messages
-
-case class Request(query: Option[String])
