--- src/main/scala/com/github/scalalab3/logs/services/messages/RequestStorage.scala
@@ -1,5 +1,0 @@
-package com.github.scalalab3.logs.services.messages
-
-import akka.actor.ActorRef
-
-case class RequestStorage(req: Request, actorRef: ActorRef)
