+++ src/main/scala/com/github/scalalab3/logs/services/query/QueryServiceActor.scala
@@ -1,20 +1,19 @@
 package com.github.scalalab3.logs.services.query
 
-import akka.actor.Actor
 import com.github.scalalab3.logs.parser.QueryParserImpl._
-import com.github.scalalab3.logs.services.messages._
+import com.github.scalalab3.logs.services.{BadRequest, Request, RequestStorage, AbstractService}
 
 import scala.util.{Failure, Success}
 
-class QueryServiceActor extends Actor {
+class QueryServiceActor extends AbstractService {
 
   override def receive: Receive = {
     case RequestStorage(Request(string), storage) =>
       string.map(parse) match {
         case None => sender ! BadRequest("Empty query")
-        case Some(query) => query match {
-          case Success(q) => storage ! QuerySender(q, sender)
-          case Failure(e) => sender ! BadRequest(e.getMessage)
+        case Some(queryTry) => queryTry match {
+          case Success(query) => storage forward query
+          case Failure(error) => sender ! BadRequest(error.getMessage)
         }
       }
   }
