+++ src/main/scala/com/github/scalalab3/logs/services/query/QueryServiceRoute.scala
@@ -4,22 +4,20 @@ import akka.actor.Props
 import akka.pattern.ask
 import akka.util.Timeout
 import com.github.scalalab3.logs.json.LogJsonImplicits._
-import com.github.scalalab3.logs.services.messages._
-import com.github.scalalab3.logs.services.{ActorCreationSupport, StorageProvider}
+import com.github.scalalab3.logs.services._
 import spray.http.StatusCodes
 import spray.httpx.marshalling.ToResponseMarshallable
 import spray.routing.HttpService
 
-import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
 
-trait QueryServiceRoute extends HttpService with ActorCreationSupport {
+trait QueryServiceRoute extends HttpService {
   self: StorageProvider =>
 
   implicit val timeout = Timeout(5.seconds)
-  implicit def exc: ExecutionContext
+  implicit def executionContext = actorRefFactory.dispatcher
 
-  val queryActor = create(Props[QueryServiceActor], "query-actor")
+  val queryActor = actorRefFactory.actorOf(Props[QueryServiceActor], "query-actor")
 
   val queryRoute = get {
     (path("query") & parameter('query.?)) { string =>
