+++ src/test/scala/com/github/scalalab3/logs/services/query/QueryServiceTest.scala
@@ -1,37 +1,24 @@
 package com.github.scalalab3.logs.services.query
 
-import akka.actor.{ActorRef, ActorRefFactory, Props}
-import akka.testkit.{TestActor, TestProbe}
-import com.github.scalalab3.logs.services.messages.{LogsResponse, RequestStorage}
-import com.github.scalalab3.logs.services.{ActorCreationSupport, StorageProvider}
+import akka.actor.{Actor, ActorRef, ActorRefFactory}
+import akka.testkit.{TestActorRef, TestProbe}
+import com.github.scalalab3.logs.services.{LogsResponse, RequestStorage, StorageProvider}
 import com.github.scalalab3.logs.tests.{GenLog, LogJsonSpecification, Specs2RouteTest}
-import org.specs2.specification.AfterAll
 import spray.http.{HttpMethods, StatusCodes, Uri}
 import spray.routing.MethodRejection
 
-import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.DurationInt
+class QueryServiceTest extends LogJsonSpecification with Specs2RouteTest {
 
-class QueryServiceTest extends LogJsonSpecification with AfterAll with Specs2RouteTest {
-
-  implicit val routeTestTimeout = RouteTestTimeout(5.second)
-
-  val queryService = TestProbe()
-  queryService.setAutoPilot {
-    new TestActor.AutoPilot {
-      def run(sender: ActorRef, msg: Any) = msg match {
-        case RequestStorage(_, _) => sender ! LogsResponse(Seq(log))
-          TestActor.KeepRunning
-        case _ => TestActor.NoAutoPilot
-      }
+  val queryTestService = TestActorRef(new Actor {
+    def receive = {
+      case RequestStorage(_, _) => sender ! LogsResponse(Seq(log))
     }
-  }
+  })
 
-  val subject = new QueryServiceRoute with StorageProvider with ActorCreationSupport {
-    override implicit def exc: ExecutionContext = system.dispatcher
+  val subject = new QueryServiceRoute with StorageProvider {
     override implicit def actorRefFactory: ActorRefFactory = system
     override def dbService: ActorRef = TestProbe().ref
-    override def create(props: Props, name: String): ActorRef = queryService.ref
+    override val queryActor: ActorRef = queryTestService
   }
 
   val queryRoute = subject.queryRoute
