+++ tests/src/main/scala/com/github/scalalab3/logs/tests/Specs2RouteTest.scala
@@ -3,6 +3,9 @@ package com.github.scalalab3.logs.tests
 import org.specs2.specification.AfterAll
 import spray.testkit.RouteTest
 
-trait Specs2RouteTest extends RouteTest with Specs2Interface  with AfterAll {
+import scala.concurrent.duration.DurationInt
+
+trait Specs2RouteTest extends RouteTest with Specs2Interface with AfterAll {
+  implicit val routeTestTimeout = RouteTestTimeout(5.second)
   def afterAll() = system.terminate
 }
